+++ ignite-reactjs-2023/03-ignite-timer/src/contexts/CyclesContext.tsx
@@ -1,0 +1,101 @@
+import { createContext, ReactNode, useState } from "react";
+
+interface CreateCycleData{
+    task: string;
+    minutesAmount: number;
+}
+
+interface Cycle {
+    id: string;
+    task: string;
+    minutesAmount: number;
+    startDate: Date;
+    interruptedDate?: Date;
+    finishedDate?: Date;
+}
+
+interface CycleContextType {
+    activeCycle: Cycle | undefined;
+    activeCycleId: string | null;
+    amountSecondsPassed: number;
+    markCurrentCycleAsFinished: () => void;
+    setSecondsPassed: (seconds: number) => void;
+    createNewCycle: (data: CreateCycleData) => void;
+    interruptCurrentCycle: () => void;
+}
+
+export const CyclesContext = createContext({} as CycleContextType);
+
+interface CyclesContextProviderProps {
+    children: ReactNode;
+}
+
+export function CyclesContextProvider({ children }: CyclesContextProviderProps){
+    const [cycles, setCycles] = useState<Cycle[]>([]);
+    const [activeCycleId, setActiveCycleId] = useState<string | null>(null);
+    const [amountSecondsPassed, setAmountSecondsPassed] = useState(0);
+
+    const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId);
+    
+    function setSecondsPassed(seconds: number){
+        setAmountSecondsPassed(seconds);
+    }
+
+    function markCurrentCycleAsFinished(){
+        setCycles((state) =>
+            state.map((cycle) => {
+                if(cycle.id === activeCycleId){
+                    return {...cycle, finishedDate: new Date() };
+                } else {
+                    return cycle;
+                }
+            }),
+        )
+    }
+
+    function createNewCycle(data: CreateCycleData){
+        const id = String(new Date().getTime());
+        const newCycle: Cycle = {
+            id,
+            task: data.task,
+            minutesAmount: data.minutesAmount,
+            startDate: new Date(),
+        };
+        //sempre que um mudançá de estado dependeer do valor anterior, usar a notacão abaixo
+        setCycles((state) => [...state, newCycle]);
+
+        setActiveCycleId(id);
+        setAmountSecondsPassed(0);
+
+        //a função abaixo retorna os campos para o defaultValues definidos no resolver
+        //reset();
+    }
+
+    function interruptCurrentCycle(){
+        
+        setCycles(state => state.map(cycle => {
+            if(cycle.id === activeCycleId){
+                return { ...cycle, interruptedDate: new Date()}
+            }else{
+                return cycle;
+            }
+        }),
+        )
+        setActiveCycleId(null);
+    }
+
+    return (
+        <CyclesContext.Provider
+            value={{
+            activeCycle,
+            activeCycleId,
+            markCurrentCycleAsFinished, 
+            amountSecondsPassed, 
+            setSecondsPassed,
+            createNewCycle,
+            interruptCurrentCycle
+        }}>
+            {children}
+        </CyclesContext.Provider>
+    );
+}
