+++ ignite-reactjs-2023/03-ignite-timer/src/pages/Home/components/NewCycleForm/index.tsx
@@ -1,7 +1,7 @@
 import { useForm, useFormContext } from "react-hook-form";
 import { FormContainer, MinutesAmountInput, TaskInput } from "./styles";
 import { useContext } from "react";
-import { CyclesContext } from "../..";
+import { CyclesContext } from "../../../../contexts/CyclesContext";
 
 export function NewCycleForm(){
     const { activeCycle } = useContext(CyclesContext);
