+++ ignite-reactjs-2023/03-ignite-timer/src/pages/Home/index.tsx
@@ -5,29 +5,10 @@ import { HomeContainer,
          StopCountdownButton} from "./styles";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as zod from "zod";
-import { createContext, useEffect, useState } from "react";
-import { differenceInSeconds } from "date-fns";
+import { useContext } from "react";
 import { NewCycleForm } from "./components/NewCycleForm";
 import { Countdown } from "./components/Countdown";
-
-interface Cycle {
-    id: string;
-    task: string;
-    minutesAmount: number;
-    startDate: Date;
-    interruptedDate?: Date;
-    finishedDate?: Date;
-}
-
-interface CycleContextType {
-    activeCycle: Cycle | undefined;
-    activeCycleId: string | null;
-    amountSecondsPassed: number;
-    markCurrentCycleAsFinished: () => void;
-    setSecondsPassed: (seconds: number) => void;
-}
-
-export const CyclesContext = createContext({} as CycleContextType);
+import { CyclesContext } from "../../contexts/CyclesContext";
 
 const newCycleFormValidationSchema = zod.object({
     task: zod.string().min(1, "Informe a tarefa"),
@@ -47,9 +28,7 @@ const newCycleFormValidationSchema = zod.object({
 type NewCycleFormData = zod.infer<typeof newCycleFormValidationSchema>;
 
 export function Home(){
-    const [cycles, setCycles] = useState<Cycle[]>([]);
-    const [activeCycleId, setActiveCycleId] = useState<string | null>(null);
-    const [amountSecondsPassed, setAmountSecondsPassed] = useState(0);
+    const {activeCycle, createNewCycle, interruptCurrentCycle } = useContext(CyclesContext);
 
     const newCycleForm = useForm<NewCycleFormData>({
         resolver: zodResolver(newCycleFormValidationSchema),
@@ -59,73 +38,22 @@ export function Home(){
         }
     });
 
-    const { handleSubmit, watch, reset } = newCycleForm
-
-    const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId);
-
-    function setSecondsPassed(seconds: number){
-        setAmountSecondsPassed(seconds);
-    }
-
-    function markCurrentCycleAsFinished(){
-        setCycles((state) =>
-            state.map((cycle) => {
-                if(cycle.id === activeCycleId){
-                    return {...cycle, finishedDate: new Date() };
-                } else {
-                    return cycle;
-                }
-            }),
-        )
-    }
-
-    function handleCreateNewCycle(data: NewCycleFormData){
-        const id = String(new Date().getTime());
-        const newCycle: Cycle = {
-            id,
-            task: data.task,
-            minutesAmount: data.minutesAmount,
-            startDate: new Date(),
-        };
-        //sempre que um mudançá de estado dependeer do valor anterior, usar a notacão abaixo
-        setCycles((state) => [...state, newCycle]);
-
-        setActiveCycleId(id);
-        setAmountSecondsPassed(0);
-
-        //a função abaixo retorna os campos para o defaultValues definidos no resolver
-        reset();
-    }
-
-    function handleInterruptCycle(){
-        
-        setCycles(state => state.map(cycle => {
-            if(cycle.id === activeCycleId){
-                return { ...cycle, interruptedDate: new Date()}
-            }else{
-                return cycle;
-            }
-        }),
-        )
-        setActiveCycleId(null);
-    }
+    const { handleSubmit, watch, reset } = newCycleForm;
 
     const task = watch("task");
     const isSubmitDisabled = !task;
 
     return (
         <HomeContainer>
-            <form onSubmit={handleSubmit(handleCreateNewCycle)} action="">
-                <CyclesContext.Provider
-                    value={{ activeCycle, activeCycleId, markCurrentCycleAsFinished, amountSecondsPassed, setSecondsPassed }}>
-                    <FormProvider {...newCycleForm}>
-                        <NewCycleForm />
-                    </FormProvider>
-                    <Countdown />
-                </CyclesContext.Provider>
+            <form onSubmit={handleSubmit(createNewCycle)} action="">
+
+                <FormProvider {...newCycleForm}>
+                    <NewCycleForm />
+                </FormProvider>
+                <Countdown />
 
                 { activeCycle ?  (
-                    <StopCountdownButton onClick={handleInterruptCycle} type="submit">
+                    <StopCountdownButton onClick={interruptCurrentCycle} type="submit">
                         <HandPalm size={24} />
                         Interromper
                     </StopCountdownButton>
